bugfixes
[X] streamlined functions                                 [X] fixed false positive (O) when two guess chars in places before the one goal char
[X] choice instead of randint
[X] input that's not in word list now counts as a try
